requirements.txt
flask
google-generativeai
gTTS
google-api-python-client
google-auth-oauthlib
google-cloud-secret-manager
google-cloud-firestore
sox
